test(zed): isolate stream boundary fixture
